refactor: moved auth components in auth folder
