+++ sea-of-code/src/pages/auth/components/AuthHeader.tsx
