+++ sea-of-code/src/pages/auth/components/RegisterLink.tsx
